# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentation",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api/v1",
            "description": "Local Development"
        },
        {
            "url": "/api/v1"
        }
    ],
    "paths": {
...
        "/Admin/Logs/Activities/Storage-Stats": {
            "post": {
                "tags": [
                    "Activity Logs"
                ],
                "summary": "Get activity log storage statistics",
                "description": "Retrieve storage usage statistics for activity logs including table size, row count, retention info, and recommendations",
                "operationId": "862cc34e21e19c7f892c89c33030253a",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "statusCode": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "data": {
                                            "properties": {
                                                "total_records": {
                                                    "type": "integer",
                                                    "example": 50000
                                                },
                                                "table_size_mb": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 125.5
                                                },
                                                "data_size_mb": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 80.3
                                                },
                                                "indexes_size_mb": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 45.2
                                                },
                                                "oldest_record": {
                                                    "type": "string",
                                                    "example": "2024-01-15 10:30:00"
                                                },
                                                "newest_record": {
                                                    "type": "string",
                                                    "example": "2025-02-27 15:30:00"
                                                },
                                                "records_last_30_days": {
                                                    "type": "integer",
                                                    "example": 15000
                                                },
                                                "records_last_7_days": {
                                                    "type": "integer",
                                                    "example": 3500
                                                },
                                                "avg_records_per_day": {
                                                    "type": "integer",
                                                    "example": 500
                                                },
                                                "projected_records_in_90_days": {
                                                    "type": "integer",
                                                    "example": 45000
                                                },
                                                "projected_size_in_90_days_mb": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 250.5
                                                },
                                                "retention_policy": {
                                                    "properties": {
                                                        "lama_hari_penyimpanan": {
                                                            "type": "integer",
                                                            "example": 90
                                                        },
                                                        "pembersihan_otomatis_diaktifkan": {
                                                            "type": "boolean",
                                                            "example": true
                                                        },
                                                        "pembersihan_berikutnya": {
                                                            "type": "string",
                                                            "example": "2025-03-02 02:00:00"
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "recommendations": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string"
                                                    },
                                                    "example": [
                                                        "Pertimbangkan untuk mengurangi masa penyimpanan dari 90 menjadi 30-60 hari"
                                                    ]
                                                },
                                                "warnings": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string"
                                                    }
                                                },
                                                "health_status": {
                                                    "type": "string",
                                                    "enum": [
                                                        "healthy",
                                                        "warning"
                                                    ],
                                                    "example": "healthy"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/401"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "responses": {
...
            "401": {
                "description": "Forbidden: Unauthenticated",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "example": "Unauthenticated"
                                },
                                "status": {
                                    "type": "string",
                                    "example": "error"
                                },
                                "statusCode": {
                                    "type": "integer",
                                    "example": 401
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}